+++ src/Resources/contao/languages/de/tl_internetschach_spieler.php
@@ -12,11 +12,13 @@
 
 // Standardfunktionen
 $GLOBALS['TL_LANG']['tl_internetschach_spieler']['new'] = array('Neuer Spieler', 'Neuen Spieler anlegen');
-$GLOBALS['TL_LANG']['tl_internetschach_spieler']['edit'] = array('Spieler %s bearbeiten', 'Spieler %s bearbeiten');
-$GLOBALS['TL_LANG']['tl_internetschach_spieler']['copy'] = array('Spieler %s kopieren', 'Spieler %s kopieren');
-$GLOBALS['TL_LANG']['tl_internetschach_spieler']['delete'] = array('Spieler %s löschen', 'Spieler %s löschen');
-$GLOBALS['TL_LANG']['tl_internetschach_spieler']['toggle'] = array('Spieler %s aktivieren/deaktivieren', 'Spieler %s aktivieren/deaktivieren');
-$GLOBALS['TL_LANG']['tl_internetschach_spieler']['show'] = array('Details zum Spieler %s anzeigen', 'Details zum Spieler %s anzeigen');
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['edit'] = array("Spieler %s bearbeiten", "Spieler %s bearbeiten");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['copy'] = array("Spieler %s kopieren", "Spieler %s kopieren");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['delete'] = array("Spieler %s löschen", "Spieler %s löschen");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['toggle'] = array("Spieler %s aktivieren/deaktivieren", "Spieler %s aktivieren/deaktivieren");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['show'] = array("Details zum Spieler %s anzeigen", "Details zum Spieler %s anzeigen");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['copyToAnmeldung'] = array("Spieler %s zu den Anmeldungen kopieren", "Spieler %s zu den Anmeldungen kopieren");
+$GLOBALS['TL_LANG']['tl_internetschach_spieler']['copyToAnmeldung_confirm'] = "Wollen Sie wirklich den Spieler %s zu den Anmeldungen kopieren?";
 
 // Formularfelder
 $GLOBALS['TL_LANG']['tl_internetschach_spieler']['name_legend'] = 'Name';